+++ CHANGELOG.md
@@ -122,6 +122,16 @@
   - `drop`
   - `invert`
   - `duplicated`
+  - `groupby.agg`
+  - `groupby.min`
+  - `groupby.max`
+  - `groupby.count`
+  - `groupby.sum`
+  - `groupby.mean`
+  - `groupby.median`
+  - `groupby.std`
+  - `groupby.var`
+
 - Reuse row count from the relaxed query compiler in `get_axis_len`.
 
 #### Bug Fixes